+++ src/main/java/com/miracle/coordifit/fitting/service/IFittingAnalysisService.java
@@ -1,0 +1,8 @@
+package com.miracle.coordifit.fitting.service;
+
+import com.miracle.coordifit.fitting.dto.FittingAnalysisRequest;
+import com.miracle.coordifit.fitting.dto.FittingAnalysisResponse;
+
+public interface IFittingAnalysisService {
+	FittingAnalysisResponse analyze(FittingAnalysisRequest request);
+}
